day 8 part 2(this one was not made well)

+++ c#/day 8 part 1/Program.cs
@@ -30,4 +30,4 @@ while(currentNode != "ZZZ")
     result++;
 }
 
-Console.WriteLine(result);
+Console.WriteLine((long)result);
